+++ .vscode/tasks.json
@@ -7,7 +7,7 @@
             "type": "process",
             "args": [
                 "build",
-                "${workspaceFolder}/src/WorkloadManager/Monai.Deploy.WorkloadManager.csproj",
+                "${workspaceFolder}/src/WorkflowManager/Monai.Deploy.WorkflowManager.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
@@ -19,7 +19,7 @@
             "type": "process",
             "args": [
                 "publish",
-                "${workspaceFolder}/src/WorkloadManager/Monai.Deploy.WorkloadManager.csproj",
+                "${workspaceFolder}/src/WorkflowManager/Monai.Deploy.WorkflowManager.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
@@ -32,7 +32,7 @@
             "args": [
                 "watch",
                 "run",
-                "${workspaceFolder}/src/WorkloadManager/Monai.Deploy.WorkloadManager.csproj",
+                "${workspaceFolder}/src/WorkflowManager/Monai.Deploy.WorkflowManager.csproj",
                 "/property:GenerateFullPaths=true",
                 "/consoleloggerparameters:NoSummary"
             ],
