+++ src/CLI/Test/Monai.Deploy.WorkflowManager.CLI.Test.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
-    <RootNamespace>Monai.Deploy.WorkloadManager.CLI.Test</RootNamespace>
+    <RootNamespace>Monai.Deploy.WorkflowManager.CLI.Test</RootNamespace>
 
     <IsPackable>false</IsPackable>
   </PropertyGroup>
