+++ src/Configuration/WorkloadManagerOptions.cs
@@ -1,16 +1,16 @@
 ﻿// SPDX-FileCopyrightText: © 2021-2022 MONAI Consortium
 // SPDX-License-Identifier: Apache License 2.0
 
-namespace Monai.Deploy.WorkloadManager.Configuration
+namespace Monai.Deploy.WorkflowManager.Configuration
 {
-    public class WorkloadManagerOptions
+    public class WorkflowManagerOptions
     {
         /// <summary>
         /// Name of the key for retrieve database connection string.
         /// </summary>
-        public const string DatabaseConnectionStringKey = "WorkloadManagerDatabase";
+        public const string DatabaseConnectionStringKey = "WorkflowManagerDatabase";
 
-        public WorkloadManagerOptions()
+        public WorkflowManagerOptions()
         {
 
         }
