+++ src/Contracts/Rest/ServiceStatus.cs
@@ -1,7 +1,7 @@
 ﻿// SPDX-FileCopyrightText: © 2021-2022 MONAI Consortium
 // SPDX-License-Identifier: Apache License 2.0
 
-namespace Monai.Deploy.WorkloadManager.Contracts.Rest
+namespace Monai.Deploy.WorkflowManager.Contracts.Rest
 {
     /// <summary>
     /// Defines the state of a running service.
