+++ src/Database/Monai.Deploy.WorkflowManager.Database.csproj
@@ -2,13 +2,9 @@
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
-    <RootNamespace>Monai.Deploy.WorkloadManager.Configuration</RootNamespace>
+    <RootNamespace>Monai.Deploy.WorkflowManager.Database</RootNamespace>
     <CodeAnalysisRuleSet>..\.sonarlint\project-monai_monai-deploy-workflow-managercsharp.ruleset</CodeAnalysisRuleSet>
   </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Options" Version="6.0.0" />
-  </ItemGroup>
 
   <ItemGroup>
     <Compile Remove="Test\**" />
