+++ src/WorkflowManager/Monai.Deploy.WorkflowManager.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net6.0</TargetFramework>
-    <RootNamespace>Monai.Deploy.WorkloadManager</RootNamespace>
+    <RootNamespace>Monai.Deploy.WorkflowManager</RootNamespace>
     <CodeAnalysisRuleSet>..\.sonarlint\project-monai_monai-deploy-workflow-managercsharp.ruleset</CodeAnalysisRuleSet>
   </PropertyGroup>
 
@@ -38,9 +38,9 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Configuration\Monai.Deploy.WorkloadManager.Configuration.csproj" />
-    <ProjectReference Include="..\Contracts\Monai.Deploy.WorkloadManager.Contracts.csproj" />
-    <ProjectReference Include="..\Database\Monai.Deploy.WorkloadManager.Database.csproj" />
+    <ProjectReference Include="..\Configuration\Monai.Deploy.WorkflowManager.Configuration.csproj" />
+    <ProjectReference Include="..\Contracts\Monai.Deploy.WorkflowManager.Contracts.csproj" />
+    <ProjectReference Include="..\Database\Monai.Deploy.WorkflowManager.Database.csproj" />
   </ItemGroup>
 
   <ItemGroup>
