+++ src/WorkflowManager/Services/Common/IMonaiService.cs
@@ -1,9 +1,9 @@
-// SPDX-FileCopyrightText: © 2021-2022 MONAI Consortium
+// SPDX-FileCopyrightText: ï¿½ 2021-2022 MONAI Consortium
 // SPDX-License-Identifier: Apache License 2.0
 
-using Monai.Deploy.WorkloadManager.Contracts.Rest;
+using Monai.Deploy.WorkflowManager.Contracts.Rest;
 
-namespace Monai.Deploy.WorkloadManager.Services.Common
+namespace Monai.Deploy.WorkflowManager.Services.Common
 {
     public interface IMonaiService
     {
