+++ src/WorkflowManager/appsettings.Development.json
@@ -1,8 +1,8 @@
 {
   "ConnectionStrings": {
-    "WorkloadManagerDatabase": "Data Source=mwmdev.db"
+    "WorkflowManagerDatabase": "Data Source=mwmdev.db"
   },
-  "WorkloadManager": {
+  "WorkflowManager": {
   },
   "Logging": {
     "LogLevel": {
