+++ src/WorkflowManager/appsettings.json
@@ -1,8 +1,8 @@
 {
   "ConnectionStrings": {
-    "WorkloadManagerDatabase": "Data Source=mwm.db"
+    "WorkflowManagerDatabase": "Data Source=mwm.db"
   },
-  "WorkloadManager": {
+  "WorkflowManager": {
   },
   "Logging": {
     "LogLevel": {
@@ -32,7 +32,7 @@
       "MaxFileSize": 10485760,
       "IncludeScopes": true,
       "MaxQueueSize": 100,
-      "TextBuilderType": "Monai.Deploy.WorkloadManager.Logging.FileLoggingTextFormatter, Monai.Deploy.WorkloadManager",
+      "TextBuilderType": "Monai.Deploy.WorkflowManager.Logging.FileLoggingTextFormatter, Monai.Deploy.WorkflowManager",
       "Files": [
         {
           "Path": "MWM-<date>-<counter>.log"
