+++ src/run-tests.sh
@@ -42,8 +42,8 @@ mkdir -p "$RESULTS_DIR"
 
 echo "##### Building MONAI Deploy Workload Manager..."
 cd $TOP/src
-dotnet build Monai.Deploy.WorkloadManager.sln
+dotnet build Monai.Deploy.WorkflowManager.sln
 
 echo "Executing all tests"
-dotnet test -v=$VERBOSITY --runtime linux-x64 --results-directory "$RESULTS_DIR" --collect:"XPlat Code Coverage" --settings "$SCRIPT_DIR/coverlet.runsettings" Monai.Deploy.WorkloadManager.sln
+dotnet test -v=$VERBOSITY --runtime linux-x64 --results-directory "$RESULTS_DIR" --collect:"XPlat Code Coverage" --settings "$SCRIPT_DIR/coverlet.runsettings" Monai.Deploy.WorkflowManager.sln
 exit $?
